+++ test/modes.js
@@ -30,7 +30,9 @@ describe('modes', function() {
         'ns': 'ns',
         'ns$uri': 'http://ns',
         'xmlns': 'ns',
-        'xmlns$uri': 'http://ns'
+        'xmlns$uri': 'http://ns',
+        'xsi': 'xsi',
+        'xsi$uri': 'http://www.w3.org/2001/XMLSchema-instance'
       });
 
       assert.equal(decodeEntities(el.attrs.foo), '"');